Changed Room::exits into a std::set, split the utils file
