+++ include/DungeonManager/Room.hpp
@@ -1,6 +1,13 @@
 #pragma once
 #include <SFML/Graphics.hpp>
-#include <vector>
+#include <set>
+
+struct Vector2iLess
+{
+    bool operator()(const sf::Vector2i& a, const sf::Vector2i& b) const;
+};
+
+using Vector2iSet = std::set<sf::Vector2i, Vector2iLess>;
 
 class Room
 {
@@ -13,9 +20,9 @@ class Room
         Trap,
         Trader
     };
-    std::vector<sf::Vector2i> exits;
+    Vector2iSet exits;
     RoomType type;
-    Room(Room::RoomType type = Room::RoomType::Empty, std::vector<sf::Vector2i> exits = {});
+    Room(Room::RoomType type = Room::RoomType::Empty, Vector2iSet exits = {});
     virtual ~Room() = default;
 };
 
@@ -24,5 +31,5 @@ class Corridor : public Room
     public:
     bool vertical;
     bool intersection;
-    Corridor(bool is_vertical, Room::RoomType type = Room::RoomType::Empty, std::vector<sf::Vector2i> exits = {});
+    Corridor(bool is_vertical, Room::RoomType type = Room::RoomType::Empty, Vector2iSet exits = {});
 };
